Split up architecture module into subpackage
